requirements.txt
python-telegram-bot==20.6
pandas
ccxt
ta
flask
websockets
